refactor codeql versions on single container
