+++ SAST/codeql/codeql_v2_13_1/codeql.py
@@ -2,9 +2,9 @@ from pathlib import Path
 import yaml
 from typing import Dict
 import sys
-sys.path.append('/tp-framework/SAST/codeql/core')
+sys.path.append('/tp-framework/SAST')
 
-from codeql_core import CodeQL
+from codeql.core.codeql import CodeQL
 
 class CodeQL_v_2_13_1(CodeQL):
